+++ packages/@react-elf-types/button/src/index.d.ts
@@ -28,6 +28,8 @@ type ButtonSize = "extra-small" | "small" | "default" | "large" | "extra-large";
 
 type ButtonShape = "square" | "round" | "circle";
 
+type OrientationType = "horizontal" | "vertical";
+
 interface ButtonProps {
   children?: ReactNode;
   variant?: ButtonVariant;