+++ packages/@react-elf/dialog/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @react-elf/dialog
 
+## 0.0.56
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @react-elf/button@0.0.56
+  - @react-elf-types/dialog@0.0.56
+  - @react-elf/shared@0.0.56
+
 ## 0.0.40
 
 ### Patch Changes